requirements.txt
python-bale-bot
mysql-connector-python~=8.0.27